rectif bouton supprimer article

+++ frontend/src/gql/posts/deletePost.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const DELETE_POST = gql`
-    mutation Mutation($id: Float!) {
+    mutation Mutation($id: Int!) {
     deletePost(id: $id)
     }`
 ;
